simple map
R script for producing a simple color map of East Africa. it can be easily modified for any region at the desired spatial scale and can be run as is (if the necessary packages are installed).  Maybe, YOU, John Q. Public, will find it  useful. Is there anybody out there?